move subCommand to dedicated dir and add todo
