+++ src/Core/Domain/Product/Command/SubCommand/UpdateProductSubCommandInterface.php
@@ -24,7 +24,7 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  */
 
-namespace PrestaShop\PrestaShop\Core\Domain\Product\Command;
+namespace PrestaShop\PrestaShop\Core\Domain\Product\Command\SubCommand;
 
 interface UpdateProductSubCommandInterface
 {
